+++ lib/model/middle-layer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @milaboratories/pl-model-middle-layer
 
+## 1.8.13
+
+### Patch Changes
+
+- 9508f78: PFrames module interface declared
+
 ## 1.8.12
 
 ### Patch Changes